+++ compositions/useLocalStorage.ts
@@ -1,7 +1,6 @@
 import { onMounted } from '@vue/composition-api';
 /* Types */
-import { TypedVuexStore } from '@/store';
-import { HistoryItem } from '@/store/history';
+import { TypedVuexStore, HistoryItem } from '~/types';
 
 const getHistoryFromLocalStorage = (): Array<any> => {
   let storage: Array<any> | string | boolean = false;
